pyproject.toml
[project]
name = "arxiv-rag"
version = "0.1.0"
description = "A simple RAG model for arXiv papers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "llama-index>=0.12.8",
    "llama-index-embeddings-fastembed>=0.3.0",
    "llama-index-llms-groq>=0.3.1",
    "llama-index-readers-file>=0.4.1",
    "llama-index-vector-stores-pinecone>=0.4.2",
    "python-dotenv>=1.0.1",
]
